requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
faiss-cpu
langchain_google_genai
